Fix false negative in integration tests
The integration tests would fail when the report cannot be saved
into the root directory.  This is, however, expected when the tests
run locally as non-root.

+++ tests-py3/scripts/run-integration-test.py
@@ -69,11 +69,17 @@ def main(args):
     try:
         return _execute_as_site_user(site, args)
     finally:
+        if not _is_dockerized():
+            return
         if os.path.exists("/results"):
             shutil.rmtree("/results")
             os.mkdir("/results")
         shutil.copy(site.path("junit.xml"), "/results")
         shutil.copytree(site.path("var/log"), "/results/logs")
+
+
+def _is_dockerized():
+    return Path("/.dockerenv").exists()
 
 
 def _centos6_mtab_fix():
